requirements.txt
fastapi
uvicorn
pandas
requests
beautifulsoup4
sentence-transformers
faiss-cpu
rank_bm25
uvloop; sys_platform == 'linux'
numpy